# api/swagger/swagger.json
{
  "info": {
    "title": "Location Service",
    "description": "This service is responsible for delivering messages to / from applications",
    "version": "1.0.0"
  },
  "swagger": "2.0",
  "paths": {
    "/geofences": {
      "post": {
        "description": "Save GeoFence",
        "operationId": "saveGeoFence",
        "tags": [
          "geoFence"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "geoFence",
            "description": "GeoFence",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/GeoFence"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "X-Tenant-Id",
            "description": "Tenant Id",
            "type": "string",
            "in": "header",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "201": {
            "description": "GeoFence",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/GeoFence"
            }
          },
          "409": {
            "description": "Conflict",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Error",
            "type": "object",
            "schema": null
          }
        }
      }
    },
...
    "/georecordings": {
      "post": {
        "description": "Save GeoRecording",
        "operationId": "saveGeoRecording",
        "tags": [
          "geoRecording"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "geoRecording",
            "description": "GeoRecording",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/GeoRecording"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "X-Tenant-Id",
            "description": "Tenant Id",
            "type": "string",
            "in": "header",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "201": {
            "description": "GeoRecording",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/GeoRecording"
            }
          },
          "409": {
            "description": "Conflict",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Error",
            "type": "object",
            "schema": null
          }
        }
      },
      "get": {
        "description": "Page GeoRecordings",
        "operationId": "page",
        "tags": [
          "geoRecording"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "query",
            "description": "Query String - default '{}'",
            "type": "string",
            "in": "query",
            "required": false
          },
          {
            "name": "limit",
            "description": "Limit - default 10",
            "in": "query",
            "required": false
          },
          {
            "name": "offset",
            "description": "Offset - default 0",
            "in": "query",
            "required": false
          },
          {
            "name": "X-Tenant-Id",
            "description": "Tenant Id",
            "type": "string",
            "in": "header",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "PageResult",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/PageResult"
            }
          },
          "500": {
            "description": "Internal Error",
            "type": "object",
            "schema": null
          }
        }
      }
    },
    "/georecordings/{id}": {
      "put": {
        "description": "Update GeoRecording",
        "operationId": "updateGeoRecording",
        "tags": [
          "geoRecording"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Id of GeoRecording to update",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "geoRecording",
            "description": "GeoRecording",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/GeoRecording"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "X-Tenant-Id",
            "description": "Tenant Id",
            "type": "string",
            "in": "header",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "201": {
            "description": "GeoRecording",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/GeoRecording"
            }
          },
          "409": {
            "description": "Conflict",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Error",
            "type": "object",
            "schema": null
          }
        }
      }
    },
...
  },
  "definitions": {
    "GeoRecording": {
      "type": "object",
      "required": [
        "ticketId",
        "speed",
        "heading",
        "coordinates"
      ],
      "properties": {
        "ticketId": {
          "type": "string"
        },
        "speed": {
          "type": "number"
        },
        "heading": {
          "type": "number"
        },
        "coordinates": {
          "type": [
            "number"
          ]
        }
      }
    },
    "GeoFence": {
      "type": "object",
      "required": [
        "accountId",
        "coordinates"
      ],
      "properties": {
        "accountId": {
          "type": "string"
        },
        "coordinates": {
          "type": [
            [
              "number"
            ]
          ]
        }
      }
    },
    "PageResult": {
      "type": "object",
      "required": [
        "limit",
        "offset",
        "total"
      ],
      "properties": {
        "limit": {
          "type": "number"
        },
        "offset": {
          "type": "number"
        },
        "total": {
          "type": "number"
        }
      }
    },
...
    "Error": {
      "type": "object",
      "required": [
        "errorMessage"
      ],
      "properties": {
        "errorMessage": {
          "type": "string"
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}